requirements.txt
africastalking==1.2.7
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contextlib2==21.6.0
coverage==7.4.1
cryptography==3.4.6
defusedxml==0.8.0rc2
Django==4.2.10
djangorestframework==3.14.0
ecdsa==0.18.0
idna==3.6
oauthlib==3.2.2
psycopg2==2.9.9
psycopg2-binary==2.8.6
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
python-jose==3.3.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schema==0.7.5
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
